914: fix pylint
